+++ plume/src/main/kotlin/io/github/plume/oss/domain/mappers/VertexMapper.kt
@@ -54,6 +54,7 @@ object VertexMapper {
                 else -> map[it] = value as Any
             }
         }
+        val unknown = io.shiftleft.semanticcpg.language.types.structure.File.UNKNOWN()
         return when (map["label"] as String) {
             ArrayInitializer.Label() -> NewArrayInitializerBuilder()
                 .order(map[ORDER] as Int)
@@ -65,7 +66,7 @@ object VertexMapper {
                 .version(map[VERSION] as String)
             File.Label() -> NewFileBuilder()
                 .name(map[NAME] as String)
-                .hash(Option.apply(map.getOrDefault("HASH", "<unknown>") as String))
+                .hash(Option.apply(map.getOrDefault(HASH, unknown) as String))
                 .order(map[ORDER] as Int)
             Method.Label() -> NewMethodBuilder()
                 .astParentFullName(map[AST_PARENT_FULL_NAME] as String)