+++ plume/src/main/kotlin/io/github/plume/oss/drivers/JanusGraphDriver.kt
@@ -5,6 +5,7 @@ import io.shiftleft.codepropertygraph.generated.EdgeTypes
 import io.shiftleft.codepropertygraph.generated.NodeKeyNames
 import io.shiftleft.codepropertygraph.generated.NodeKeyNames.*
 import io.shiftleft.codepropertygraph.generated.NodeTypes
+import io.shiftleft.codepropertygraph.generated.nodes.NewNodeBuilder
 import org.apache.commons.configuration.PropertiesConfiguration
 import org.apache.logging.log4j.LogManager
 import org.apache.tinkerpop.gremlin.driver.Client
@@ -63,28 +64,45 @@ class JanusGraphDriver : GremlinDriver(), ISchemaSafeDriver {
      */
     fun remoteConfig(remoteConfigPath: String) = apply { config.setProperty(REMOTE_CONFIG, remoteConfigPath) }
 
-    override fun buildSchema() {
-        val propFileConfig = PropertiesConfiguration(config.getString(REMOTE_CONFIG))
-        val cluster = Cluster.open(propFileConfig.getString("gremlin.remote.driver.clusterFile"))
-        val client = cluster.connect<Client>()
-        val results = client.submit(buildSchemaPayload())
-        println(results.toList())
-        client.close()
-        cluster.close()
+    override fun updateVertexProperty(id: Long, label: String?, key: String, value: Any) {
+        if (!g.V(id).hasNext()) return
+        g.V(id).property("_$key", value).iterate()
     }
 
-    fun buildSchemaPayload(): String {
-        val schema = StringBuilder("""
+    override fun prepareVertexProperties(v: NewNodeBuilder): Map<String, Any> =
+        super.prepareVertexProperties(v).mapKeys { "_$it" }.toMap()
+
+    override fun mapVertexKeys(props: MutableMap<Any, Any>) = props.mapKeys { it.key.toString().removePrefix("_") }
+
+    override fun buildSchema() {
+        val propFileConfig = PropertiesConfiguration(config.getString(REMOTE_CONFIG))
+        var cluster: Cluster? = null
+        var client: Client? = null
+        try {
+            cluster = Cluster.open(propFileConfig.getString("gremlin.remote.driver.clusterFile"))
+            client = cluster.connect()
+            client.submit(buildSchemaPayload())
+        } finally {
+            client?.close()
+            cluster?.close()
+        }
+
+    }
+
+    override fun buildSchemaPayload(): String {
+        val schema = StringBuilder(
+            """
             graph = JanusGraphFactory.open('/etc/opt/janusgraph/janusgraph.properties')
             mgmt = graph.openManagement()
-        """.trimIndent())
+        """.trimIndent()
+        )
         schema.append("\n// Vertex labels\n")
-        NodeTypes.ALL.forEach { schema.append("mgmt.containsVertexLabel('$it') ?: mgmt.makeVertexLabel('$it').make()\n")  }
+        NodeTypes.ALL.forEach { schema.append("mgmt.containsVertexLabel('$it') ?: mgmt.makeVertexLabel('$it').make()\n") }
         schema.append("// Edge labels\n")
-        EdgeTypes.ALL.forEach { schema.append("mgmt.containsEdgeLabel('$it') ?: mgmt.makeEdgeLabel('$it').make()\n")  }
+        EdgeTypes.ALL.forEach { schema.append("mgmt.containsEdgeLabel('$it') ?: mgmt.makeEdgeLabel('$it').make()\n") }
         schema.append("// Properties\n")
         NodeKeyNames.ALL.forEach { k ->
-            schema.append("mgmt.containsPropertyKey('$k') ?: mgmt.makePropertyKey('$k')")
+            schema.append("mgmt.containsPropertyKey('_$k') ?: mgmt.makePropertyKey('_$k')")
             when {
                 ExtractorConst.BOOLEAN_TYPES.contains(k) -> schema.append(".dataType(Boolean.class)")
                 ExtractorConst.INT_TYPES.contains(k) -> schema.append(".dataType(Integer.class)")
@@ -92,13 +110,15 @@ class JanusGraphDriver : GremlinDriver(), ISchemaSafeDriver {
             }
             schema.append(".cardinality(org.janusgraph.core.Cardinality.SINGLE).make()\n")
         }
-        schema.append("""
+        schema.append(
+            """
             // Indexes
-            mgmt.getGraphIndex("byFullName") != null ?: mgmt.buildIndex("byFullName", Vertex.class).addKey(mgmt.getPropertyKey($FULL_NAME)).buildCompositeIndex()
-            mgmt.getGraphIndex("bySignature") != null ?: mgmt.buildIndex("bySignature", Vertex.class).addKey(mgmt.getPropertyKey($SIGNATURE)).buildCompositeIndex()
-            mgmt.getGraphIndex("byHash") != null ?: mgmt.buildIndex("byHash", Vertex.class).addKey(mgmt.getPropertyKey($HASH)).buildCompositeIndex()
+            mgmt.getGraphIndex("byFullName") != null ?: mgmt.buildIndex("byFullName", Vertex.class).addKey(mgmt.getPropertyKey(_$FULL_NAME)).buildCompositeIndex()
+            mgmt.getGraphIndex("bySignature") != null ?: mgmt.buildIndex("bySignature", Vertex.class).addKey(mgmt.getPropertyKey(_$SIGNATURE)).buildCompositeIndex()
+            mgmt.getGraphIndex("byHash") != null ?: mgmt.buildIndex("byHash", Vertex.class).addKey(mgmt.getPropertyKey(_$HASH)).buildCompositeIndex()
             mgmt.commit()
-        """.trimIndent())
+        """.trimIndent()
+        )
         return schema.toString()
     }